fix: correct program counter type

+++ include/state/Driver.h
@@ -14,7 +14,7 @@ class IFAdder;
 class StageSynchronizer;
 
 class Driver: public Module {
-    int program_counter;
+    unsigned long program_counter;
     bool is_new_program_counter_set;
 
     bool is_nop_asserted;
